refactor: 원격 데이터 계층 정리

+++ Rodi/Data/Remote/Practice/DTO/Response/PracticeRegisterResponseDTO.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct PracticeRegisterResponseDTO: Decodable {
-    let practiceID: Int
+    let practiceID: Int64
     let status: String?
     let visitCount: Int?
     let requiredDistanceMeters: Int?
